modif signature fonctions, ajout testFusion generique
signature fonctions => ajout const !
test fusion generique added => environ 2 * moins rapide qsort

+++ src/mem/alloc.h
@@ -7,7 +7,7 @@
 void* my_calloc(size_t taille);
 void* my_realloc(void *p, size_t taille);
 
-void** allocTab2D(int lig, int col, int tailleType);
+void** allocTab2D(int lig, int col, size_t tailleType);
 
 void liberationTab2D(void** matrice, int lig);
